Split encrypt ANTLR g4 file and visitor (#10818)
