+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/main/antlr4/imports/RQLStatement.g4
@@ -23,18 +23,6 @@ showResources
     : SHOW RESOURCES (FROM schemaName)?
     ;
 
-showEncryptRules
-    : SHOW ENCRYPT (TABLE tableRule | RULES) (FROM schemaName)?
-    ;
-
-tableRule
-    : RULE tableName
-    ;
-
 schemaName
-    : IDENTIFIER
-    ;
-
-tableName
     : IDENTIFIER
     ;
